start to move interface and remove compilation errors
