--- src/RawPOSIX/src/interface/file.rs
@@ -1,56 +1,0 @@
-use parking_lot::Mutex;
-use std::convert::TryInto;
-use std::fs::{self, canonicalize, File, OpenOptions};
-use std::os::unix::io::{AsRawFd, RawFd};
-pub use std::path::{Component as RustPathComponent, Path as RustPath, PathBuf as RustPathBuf};
-use std::sync::Arc;
-pub use std::sync::LazyLock as RustLazyGlobal;
-
-#[derive(Debug)]
-pub struct ShmFile {
-    fobj: Arc<Mutex<File>>,
-    key: i32,
-    size: usize,
-}
-
-pub fn new_shm_backing(key: i32, size: usize) -> std::io::Result<ShmFile> {
-    ShmFile::new(key, size)
-}
-
-// Mimic shared memory in Linux by creating a file backing and truncating it to the segment size
-// We can then safely unlink the file while still holding a descriptor to that segment,
-// which we can use to map shared across cages.
-impl ShmFile {
-    fn new(key: i32, size: usize) -> std::io::Result<ShmFile> {
-        // open file "shm-#id"
-        let filename = format!("{}{}", "shm-", key);
-        let f = OpenOptions::new()
-            .read(true)
-            .write(true)
-            .create(true)
-            .open(filename.clone())
-            .unwrap();
-        // truncate file to size
-        f.set_len(size as u64)?;
-        // unlink file
-        fs::remove_file(filename)?;
-        let shmfile = ShmFile {
-            fobj: Arc::new(Mutex::new(f)),
-            key,
-            size,
-        };
-
-        Ok(shmfile)
-    }
-
-    //gets the raw fd handle (integer) from a rust fileobject
-    pub fn as_fd_handle_raw_int(&self) -> i32 {
-        self.fobj.lock().as_raw_fd() as i32
-    }
-}
-
-// convert a series of big endian bytes to a size
-pub fn convert_bytes_to_size(bytes_to_write: &[u8]) -> usize {
-    let sizearray: [u8; 8] = bytes_to_write.try_into().unwrap();
-    usize::from_be_bytes(sizearray)
-}
