+++ src/cage/Cargo.toml
@@ -11,3 +11,6 @@ quick_cache = "0.6.9"
 fdtables = { path = "../fdtables" }
 once_cell = "1.18" 
 parking_lot = "0.12"
+
+[features]
+disable_signals = []
