+++ src/cage/src/memory/mod.rs
@@ -1,6 +1,6 @@
 //! This module is VMMAP specific
-pub mod mem_helper;
+pub mod memory;
 pub mod vmmap;
 
-pub use mem_helper::*;
+pub use memory::*;
 pub use vmmap::*;
